requirements.txt
pytesseract
pillow
flask
gunicorn
opencv-python